backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
tensorflow==2.14.0